requirements.txt
asgiref==3.4.1
boto3==1.20.31
botocore==1.23.31
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-braces==1.15.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.7.1
django-storages==1.12.3
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
et-xmlfile==1.1.0
fonttools==4.28.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
MarkupPy==1.14
mysqlclient==2.1.0
numpy==1.22.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.5
pdfkit==1.0.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
pydyf==0.1.2
pymongo==3.12.1
pyphen==0.11.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
tablib==3.1.0
tinycss2==1.1.1
urllib3==1.26.7
weasyprint==53.4
webencodings==0.5.1
whitenoise==5.3.0
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.9
